fixed signed/unsigned comparison problem

+++ xd/xdfs/baseset.cpp
@@ -70,7 +70,7 @@ bool XdfsBaseSet::modifyStructure(const xd::StructureModify& mod_params, bool* d
 
 
     std::vector<xd::StructureModify::Action>::const_iterator it;
-    int processed_action_count = 0;
+    size_t processed_action_count = 0;
 
     // handle delete
     for (it = mod_params.actions.cbegin(); it != mod_params.actions.cend(); ++it)
